api-gateway/requirements.txt
fastapi>=0.112.2
uvicorn>=0.30.6
httpx>=0.27.0
python-jose[cryptography]
redis>=5.0.0
python-dotenv
